+++ include/generators/components/comp_info_generator.h
@@ -16,7 +16,7 @@ void generate_follower(follower_info *info, physics_info *p_info, physics_info *
 
 void generate_destroyer(destroyer_info *info);
 
-void generate_animator(animator_info *info, int min_tex_id, int max_tex_id, int frame_duration, bool destroy_on_max);
+void generate_animator(animator_info *info, int min_tex_id, int max_tex_id, float frame_duration, bool destroy_on_max);
 
 void generate_mover(mover_info *info, physics_info *p_info);
